Revert "fix : viewCount 타입 수정"
This reverts commit fb632f21a57824a4c12a6e4ca7f98bff598ffb08.

+++ src/main/java/couch/forrest/domain/place/entity/Place.java
@@ -68,6 +68,6 @@ public class Place extends BaseTimeEntity {
     private String address;
 
     @Column(name = "view_count")
-    private Long viewCount;
+    private Integer viewCount;
 
 }
